+++ saleor/graphql/app/tests/mutations/test_app_fetch_manifest.py
@@ -30,6 +30,7 @@ mutation AppFetchManifest($manifest_url: String!){
         constraint
         satisfied
       }
+      author
       extensions{
         label
         url
@@ -684,4 +685,51 @@ def test_app_fetch_manifest_with_invalid_required_saleor_version(
     errors = content["data"]["appFetchManifest"]["errors"]
     assert len(errors) == 1
     assert errors[0]["field"] == "requiredSaleorVersion"
+    assert errors[0]["code"] == AppErrorCode.INVALID.name
+
+
+def test_app_fetch_manifest_with_author(
+    staff_api_client, app_manifest, permission_manage_apps, monkeypatch
+):
+    # given
+    app_manifest["author"] = "Acme Ltd"
+    mocked_get_response = Mock()
+    mocked_get_response.json.return_value = app_manifest
+    monkeypatch.setattr(requests, "get", Mock(return_value=mocked_get_response))
+
+    # when
+    response = staff_api_client.post_graphql(
+        APP_FETCH_MANIFEST_MUTATION,
+        variables={"manifest_url": "http://localhost:3000/manifest"},
+        permissions=[permission_manage_apps],
+    )
+
+    # then
+    content = get_graphql_content(response)
+    manifest = content["data"]["appFetchManifest"]["manifest"]
+    assert len(content["data"]["appFetchManifest"]["errors"]) == 0
+    assert manifest["author"] == app_manifest["author"]
+
+
+def test_app_fetch_manifest_with_empty_author(
+    staff_api_client, app_manifest, permission_manage_apps, monkeypatch
+):
+    # given
+    app_manifest["author"] = " "
+    mocked_get_response = Mock()
+    mocked_get_response.json.return_value = app_manifest
+    monkeypatch.setattr(requests, "get", Mock(return_value=mocked_get_response))
+
+    # when
+    response = staff_api_client.post_graphql(
+        APP_FETCH_MANIFEST_MUTATION,
+        variables={"manifest_url": "http://localhost:3000/manifest"},
+        permissions=[permission_manage_apps],
+    )
+
+    # then
+    content = get_graphql_content(response)
+    errors = content["data"]["appFetchManifest"]["errors"]
+    assert len(errors) == 1
+    assert errors[0]["field"] == "author"
     assert errors[0]["code"] == AppErrorCode.INVALID.name